// src/app/models/candidatos.models.ts
export interface Candidatos {
    tipo:number;
    idEleccion:number;
    idproceso:number;
    idcandidato:number;
    tipoDocumento: string;
    numeroDocumento: string;
    nombre: string;
    apellidos:string;
    nombreCompleto:string;
    area: string;
    localidad: string;
    urlFile: string;
    descripcion: string;
}